Update FloorPlanLayoutHome1.json
Added a "visited" boolean for each square of the grid, to be set to 1 when the Clean Sweep is on that square. At the beginning of each cycle we can set all squares' visited values to "0". As I talked about during the meeting, I believe that this is functionally the same as having the clean sweep "believe that each cell is dirty" (needs to be visited) at every cycle start.