DirtContainer test and more other testing
I created more story level test into the application.
There is a particular test in the DirtContainerTest class that it fails
which in terms of behavior it should pass but it doesn’t pass base that
I may been using an old version of JUnit test in order to test
exceptions. Please feel free to point the issue in there